Conflicts:
	megamek/src/megamek/MegaMek.java

+++ megamek/src/megamek/MegaMek.java
@@ -572,7 +572,7 @@ public class MegaMek {
                 error("directory name expected"); // $NON-NLS-1$
             }
         }
-        
+
         private void processUnitValidator() throws ParseException {
             String filename;
             if (getToken() == TOK_LITERAL) {
@@ -807,7 +807,7 @@ public class MegaMek {
                             w.write(TechConstants.getLevelDisplayableName(unit
                                     .getType()));
                             w.write(",");
-                            w.write(Float.toString(unit.getTons()));
+                            w.write(Double.toString(unit.getTons()));
                             w.write(",");
                             if (unit.isClan()) {
                                 w.write("Clan,");
